% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_character_vars.R
\name{check_character_vars}
\alias{check_character_vars}
\title{Check existence of character variables.}
\usage{
check_character_vars(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}
}
\value{
Returns the test report.
}
\description{
Check the existence of character variables in a \code{GADSdat}.
}
\examples{
# tbd

}
