% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docu_mez.R
\name{check_docu_mez}
\alias{check_docu_mez}
\title{Check documentation of data sets.}
\usage{
check_docu_mez(
  sav_path,
  pdf_path,
  post_words = 2,
  case_sensitive = FALSE,
  encoding = NULL
)
}
\arguments{
\item{sav_path}{Character vector with paths to the \code{.sav} files.}

\item{pdf_path}{Character vector with paths to the \code{.pdf} files.}

\item{post_words}{Number of words after the variable names that should be extracted from the PDF.}

\item{case_sensitive}{If \code{TRUE}, upper and lower case are differentiated for variable name matching. If \code{FALSE}, case is ignored.}

\item{encoding}{Optional: The character encoding used for reading the \code{.sav} file. The default, \code{NULL}, uses the encoding specified in the file, but sometimes this value is incorrect and it is useful to be able to override it.}
}
\value{
A \code{data.frame} with the variable names, count of mentions in the \code{pdf} (\code{count}), words after the variable names (\code{post}) and the name of the data set in which the variable occurs (\code{data_set}).
}
\description{
This is a specific version of \code{\link{check_docu}} written for the \code{MEZ} study. If a variable name can not be found in the
documentation, the variable name is split at the first numeral and \code{*} is added to the stem.
}
