% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_vars.R
\name{check_empty_vars}
\alias{check_empty_vars}
\title{Check for empty variables}
\usage{
check_empty_vars(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
A \code{data.frame} with the variable name and variable label.
If no empty variables are found, an empty \code{data.frame} is returned.
}
\description{
Identifies variables that contain only missing values. Values tagged as
missing in the metadata are treated as missing.
}
