% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id_overlap.R
\name{check_id_overlap}
\alias{check_id_overlap}
\title{Check ID overlaps across datasets}
\usage{
check_id_overlap(sav_paths, id_vars, plot = TRUE)
}
\arguments{
\item{sav_paths}{A named character vector containing paths to SPSS files.}

\item{id_vars}{Character vector of ID variables to be checked.}

\item{plot}{Logical. Should Venn diagrams be drawn?}
}
\value{
A named list of `euler` objects, returned invisibly. Entries for
  skipped ID variables are `NULL`.
}
\description{
Imports selected ID variables from SPSS files and checks their overlap
across datasets. A Venn diagram is created for each ID variable that occurs
in two to five datasets.
}
