% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdc.R
\name{check_sdc}
\alias{check_sdc}
\title{Create a Statistical Disclosure Control Report.}
\usage{
check_sdc(GADSdat, vars = eatGADS::namesGADS(GADSdat), boundary = 5)
}
\arguments{
\item{GADSdat}{tbd}

\item{vars}{tbd}

\item{boundary}{Integer number: categories with less than or equal to \code{boundary} observations will be flagged}
}
\value{
A \code{data.frame}.
}
\description{
Create a statistical disclosure control report: Which variables have categories with low absolute frequencies,
which might lead to statistical data disclosure issues?
}
\details{
Individual participants of studies such as educational large-scale assessments usually must remain non-identifiable on individual level.
This function checks all variables in a \code{GADSdat} object
for low frequency categories which might lead to statistical disclosure control issues.
Currently, only a uni-variate check is implemented.
}
\examples{
sav_path <- system.file("extdata", "LV_2011_CF.sav", package = "eatFDZ")

## don't report low frequencies for unique id variables
exclude<- c("idstud_FDZ", "idsch_FDZ")

##
sdc_report <- sdc_check(fileName = sav_path, exclude=exclude)



}
