% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_labels.R
\name{check_var_labels}
\alias{check_var_labels}
\title{Check existence of variable labels.}
\usage{
check_var_labels(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}
}
\value{
Returns the test report.
}
\description{
Check the existence of variable labels for all variables in a \code{GADSdat}.
}
\details{
Checks include
\itemize{
\item Have variable labels been assigned?
}
}
\examples{
# tbd

}
