% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_names.R
\name{check_var_names}
\alias{check_var_names}
\title{Check variable names.}
\usage{
check_var_names(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}
}
\value{
Returns the test report.
}
\description{
Check variable names of variables in a \code{GADSdat}.
}
\details{
Checks in variable names include
\itemize{
\item special signs such as \code{Umlaute}
\item column names with a \code{"."} in it.
}
}
\examples{
# tbd

}
