% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_names_length.R
\name{check_var_names_length}
\alias{check_var_names_length}
\title{Check for lengthy variable names.}
\usage{
check_var_names_length(GADSdat, boundary = 30)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{boundary}{An integer specifying the maximum length (number of characters) a variable name is allowed to have. Default is 30.}
}
\value{
A \code{data.frame} containing the variable names that exceed the boundary, along with their respective character counts.
If all variable names are within the boundary, an empty \code{data.frame} is returned.
}
\description{
This function checks whether variable names in a \code{GADSdat} object exceed the allowed number of characters.
This can help ensure that variable names adhere to naming conventions and avoid potential issues with downstream processing or compatibility.
}
\examples{
# Example usage:
# Load example GADSdat object
GADSdat <- eatGADS::import_spss(system.file("extdata", "example_data2.sav", package = "eatFDZ"))

# Check for variable names longer than 30 characters
lengthy_names <- check_var_names_length(GADSdat, boundary = 30)

# Print the result
print(lengthy_names)

}
