% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\title{Compare existing variables and meta data structure between two data sets.}
\usage{
compare_data(data1, data2, name_data1, name_data2, ID_var)
}
\arguments{
\item{data1}{Data set one, provided as a \code{GADSdat} object.}

\item{data2}{Data set one, provided as a \code{GADSdat} object.}

\item{name_data1}{In which language should the output be written? Currently only German is supported.}

\item{name_data2}{In which language should the output be written? Currently only German is supported.}

\item{ID_var}{Identifier variable in both data sets. Currently of no practical use.}
}
\description{
Compare existing variables and meta data structure between two data sets.
}
\details{
The function performs a comparison between two \code{GADSdat} objects.
Variables included in each data set are compared and reported, as well as meta data differences
(variable labels, SPSS format, value labels, missing tags).
}
\examples{
# tbd
}
