% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{FDZ Data cleaning}
\usage{
data_clean(
  fileName,
  boundary = 5,
  saveFolder = NA,
  nameListe = NULL,
  nameSyntax = NULL,
  exclude = NULL
)
}
\arguments{
\item{fileName}{Character string of the SPSS file}

\item{boundary}{Integer number: categories with less observations than [boundary] will be summarized to guarantee anonymity}

\item{saveFolder}{Character string of the target folder for SPSS syntax output. Target folder must have writing permission.}

\item{nameListe}{File name of the csv-type variable information file}

\item{nameSyntax}{File name of the SPSS syntax file}

\item{exclude}{Optional: character vector of variable which should be excluded from summarizing and transformation}
}
\value{
SPSS syntax snippet
}
\description{
Generate SPSS syntax to make individual data anonymous
}
\details{
Data from large-scale assessments often have to be non-identifiable on individual level. Function summarizes low-sized categories of polytomous variables to more general categories. Additionally, character variables are transformed into numeric factors, providing factor levels als value labels.
}
\examples{
\dontrun{
dat    <- data.frame ( foreign::read.spss ( "q:/BT2016/BT/50_Daten/03_Aufbereitet/
     06_Gesamtdatensatz/BS_LV_Primar_2016_Matchingvorlaeufig_09_erweiterteGadsversion.sav",
                      to.data.frame = FALSE, use.value.labels = FALSE, use.missings = TRUE))
classes<- sapply(dat, class)
nCat   <- sapply(dat, FUN = function ( x ) { length(unique(x))})
exclude<- intersect(which(classes=="factor"), which(nCat>90))
exclude<- colnames(dat)[exclude]
syntax <- data_clean(fileName = "q:/BT2016/BT/50_Daten/03_Aufbereitet/06_Gesamtdatensatz/
            BS_LV_Primar_2016_Matchingvorlaeufig_09_erweiterteGadsversion.sav",
            saveFolder = "N:/archiv/temp/20_fdz", nameListe = "liste2.csv",
            nameSyntax = "syntax2.txt", exclude=exclude)
}

}
