% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fdz.R
\name{download_fdz}
\alias{download_fdz}
\title{Download an empty FDZ file.}
\usage{
download_fdz(
  study = c("PISA"),
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_quest")
)
}
\arguments{
\item{study}{Name of the study}

\item{year}{Year of the assessment (only needed for longitudinal assessments).}

\item{data_type}{Type of the data.}
}
\description{
Download an empty data set from the FDZ homepage.
}
\details{
The function downloads a \code{Leerdatensatz} from the FDZ homepage.
}
\examples{
# tbd
}
