% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pisa.R
\name{download_pisa}
\alias{download_pisa}
\title{Download and import an empty PISA public use file.}
\usage{
download_pisa(
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_quest")
)
}
\arguments{
\item{year}{Year of the PISA cycle which the data is part of.}

\item{data_type}{Type of the PISA data. Currently supported is student background data (\code{"stud_quest"}).}
}
\description{
Download and import an empty PISA public use file (containing only the first data row) from the
\href{https://www.oecd.org/en/about/programmes/pisa/pisa-data.html}{OECD homepage}.
}
\details{
The function downloads a zip file from the OECD homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{\link[haven]{read_sav}}.
For downloading full PISA data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
The data is imported as a \code{GADSdat} object.
}
\examples{
\dontrun{
pisa <- download_pisa(year = "2015", data_type = "stud_quest")
}
}
