% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\docType{data}
\name{english_words}
\alias{english_words}
\title{English Language Corpus}
\format{
A character vector with 459866 entries.
}
\source{
\url{https://github.com/dwyl/english-words}
}
\usage{
english_words
}
\description{
An English language corpus downloaded from \url{https://github.com/dwyl/english-words}.
}
\keyword{datasets}
