% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\docType{data}
\name{german_words}
\alias{german_words}
\title{German Language Corpus}
\format{
A character vector with 239649 entries.
}
\source{
\url{https://github.com/davidak/wortliste}
}
\usage{
german_words
}
\description{
A German language corpus downloaded from \url{https://github.com/davidak/wortliste}.
}
\keyword{datasets}
