% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_check_report.R
\name{write_check_report}
\alias{write_check_report}
\title{Write a check report.}
\usage{
write_check_report(check_report, file_path, language = c("German", "English"))
}
\arguments{
\item{check_report}{The check report as created by \code{\link{check_all()}}.}

\item{file_path}{File destination}
}
\description{
Write a check report as created by \code{\link{check_all()}} to excel.
}
