% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Run all FDZ checks.}
\usage{
check_all(
  sav_path,
  pdf_path = NULL,
  encoding = NULL,
  missingRange = -50:-99,
  missingRegex = "missing|omitted|not reached|nicht beantwortet|ausgelassen",
  idVar = NULL,
  sdcVars = NULL
)
}
\arguments{
\item{sav_path}{Character string of the SPSS file}

\item{encoding}{Optional: The character encoding used for reading the \code{.sav} file. The default, \code{NULL}, uses the encoding specified in the file, but sometimes this value is incorrect and it is useful to be able to override it.}

\item{missingRange}{Numerical range for missing tags.}

\item{missingRegex}{Regular expression for value labels for missing tags.}

\item{idVar}{Unique identifier variable in the data set.}

\item{sdcVars}{Variable names of variables with potential statistical disclosure control issues.}
}
\value{
A \code{data.frame}.
}
\description{
Run all FDZ checks.
}
