% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Run all data checks.}
\usage{
check_all(
  sav_path,
  pdf_path = NULL,
  encoding = NULL,
  missingRange = -50:-99,
  missingRegex = "missing|omitted|not reached|nicht beantwortet|ausgelassen",
  idVar = NULL,
  sdcVars = NULL
)
}
\arguments{
\item{sav_path}{Path to the SPSS file}

\item{pdf_path}{Path to the \code{.pdf} file}

\item{encoding}{Optional: The character encoding used for reading the \code{.sav} file.
The default, \code{NULL}, uses the encoding specified in the file,
but sometimes this value is incorrect and it is useful to be able to override it.}

\item{missingRange}{Numerical range for missing tags.}

\item{missingRegex}{Regular expression for value labels for missing tags.}

\item{idVar}{Name(s) of the identifier variable in the \code{GADSdat} object. If \code{NULL}, the first variable in
the data set is taken as the \code{idVar}.}

\item{sdcVars}{Variable names of variables with potential statistical disclosure control issues.}
}
\value{
A \code{data.frame}.
}
\description{
Run all data checks.
}
\details{
This functions calls \code{\link{check_file_name}},
\code{\link{check_var_names}}, \code{\link{check_meta_encoding}},
\code{\link{check_id}}, \code{\link{check_var_labels}},
\code{\link[eatGADS]{checkMissingValLabels}},
\code{\link{check_missing_range}}, \code{\link{check_missing_regex}},
\code{\link{sdc_check}}, and \code{\link{check_docu}}.
}
\examples{
dataset <- system.file("extdata", "example_data2.sav", package = "eatFDZ")
out <- check_all(dataset)
}
