% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_character_vars.R
\name{check_character_vars}
\alias{check_character_vars}
\title{Check for the existence of character variables}
\usage{
check_character_vars(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object containing the data to be checked.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
  \item \code{varName}: The name of the character variable.
  \item \code{nUniqueValues}: The number of unique values in the character variable.
}
If no character variables exist, the function returns an empty \code{data.frame}.
}
\description{
This function identifies all character variables in a \code{GADSdat} object and counts the number of unique values in each of those variables.
Character variables are often of particular interest as they may contain free text or other non-standardized inputs
that require additional review or processing.
}
\examples{
# Example usage:

# Load example GADSdat object
GADSdat <- eatGADS::import_spss(system.file("extdata", "example_data2.sav", package = "eatFDZ"))

# Check for character variables and count unique values
character_vars_report <- check_character_vars(GADSdat)

# Print the report
print(character_vars_report)

}
