% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docu.R
\name{check_docu}
\alias{check_docu}
\title{Check documentation of a data set.}
\usage{
check_docu(sav_path, pdf_path, post_words = 2)
}
\arguments{
\item{sav_path}{Path to the \code{.sav} file.}

\item{pdf_path}{Path to the \code{.pdf} file.}

\item{post_words}{Number of words after the variable names that should be extracted from the PDF.}
}
\value{
A \code{data.frame} with the variable names, count of mentions in the PDF and words after the variable names.
}
\description{
Check if all variables of a data set saved as \code{.sav} are included in the \code{.pdf} documentation.
}
\details{
tbd
}
\examples{
# tbd

}
