% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_name.R
\name{check_file_name}
\alias{check_file_name}
\title{Check file name.}
\usage{
check_file_name(path)
}
\arguments{
\item{path}{File path.}
}
\value{
Returns the test report.
}
\description{
Check and validate a file name.
}
\details{
Checks in variable names include
\itemize{
\item special signs such as \code{Umlaute}
\item spaces
}
}
\examples{
# tbd

}
