% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Check an identifier variable.}
\usage{
check_id(GADSdat, idVar = NULL)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{idVar}{Name(s) of the identifier variable in the \code{GADSdat} object.}
}
\value{
Returns the test report.
}
\description{
Check the uniqueness and non-missingness of a variable in a \code{GADSdat}.
}
\details{
Checks include
\itemize{
\item Is the identifier variable unique (i.e., no duplicates)?
\item Are there any missing values in the identifier variable?
}
}
\examples{
# tbd

}
