% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_encoding.R
\name{check_meta_encoding}
\alias{check_meta_encoding}
\title{Check meta data encoding.}
\usage{
check_meta_encoding(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}
}
\value{
Returns the test report.
}
\description{
Check the occurence of specials signs in the meta data of a \code{GADSdat}.
}
\details{
Checks in the meta data include
\itemize{
\item special signs such as \code{Umlaute}
}
}
\examples{
# tbd

}
