% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_range.R
\name{check_missing_range}
\alias{check_missing_range}
\title{Check missing range.}
\usage{
check_missing_range(GADSdat, missingRange = -50:-99)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{missingRange}{Values which should be declared as missings.}
}
\value{
Returns the test report.
}
\description{
Check all labeled values if they fall in the range for missing values \code{GADSdat}.
}
\details{
Checks include
\itemize{
\item ...
}
}
\examples{
# tbd

}
