% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_regex.R
\name{check_missing_regex}
\alias{check_missing_regex}
\title{Check missing range.}
\usage{
check_missing_regex(
  GADSdat,
  missingRegex = "missing|omitted|not reached|nicht beantwortet|ausgelassen"
)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{missingRegex}{Value labels which should be declared as missings.}
}
\value{
Returns the test report.
}
\description{
Check all labeled values if they fall in the range for missing values \code{GADSdat}.
}
\details{
Checks include
\itemize{
\item ...
}
}
\examples{
# tbd

}
