% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdc.R
\name{check_sdc}
\alias{check_sdc}
\title{Create a Statistical Disclosure Control Report.}
\usage{
check_sdc(GADSdat, vars = eatGADS::namesGADS(GADSdat), boundary = 5)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{Character vector of variable names. Which variables should be checked?}

\item{boundary}{Integer number: categories with less than or equal to \code{boundary} observations will be flagged}
}
\value{
A \code{data.frame}.
}
\description{
Create a statistical disclosure control report: Which variables have categories with low absolute frequencies,
which might lead to statistical data disclosure issues?
}
\details{
Individual participants of studies such as educational large-scale assessments usually must remain
non-identifiable on individual level.
This function checks the specified variables in a \code{GADSdat} object
for low frequency categories which might lead to statistical disclosure control issues.
Currently, only a uni-variate check is implemented.
}
\examples{
# tbd


}
