% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_names_length.R
\name{check_var_names_length}
\alias{check_var_names_length}
\title{Check for lengthy variable names.}
\usage{
check_var_names_length(GADSdat, boundary = 30)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{boundary}{Numeric vector of lenght 1. Boundary: Number of characters variables names are allowed
to have.}
}
\value{
Returns the test report.
}
\description{
Check for lengthy variable names of variables in a \code{GADSdat}.
}
\examples{
# tbd

}
