% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\title{Compare existing variables and meta data structure between two data sets.}
\usage{
compare_data(
  data1,
  data2,
  name_data1 = "data1",
  name_data2 = "data2",
  metaExceptions = c("display_width", "labeled")
)
}
\arguments{
\item{data1}{Data set one, provided as a \code{GADSdat} object.}

\item{data2}{Data set two, provided as a \code{GADSdat} object.}

\item{name_data1}{Character vector of length 1. Name of \code{data1}.}

\item{name_data2}{Character vector of length 1. Name of \code{data2}.}

\item{metaExceptions}{Should certain meta data columns be excluded from the comparison?}
}
\description{
Compare existing variables and meta data structure between two data sets.
}
\details{
The function performs a comparison between two \code{GADSdat} objects.
Variables included in each data set are compared and reported, as well as meta data differences
(variable labels, SPSS format, value labels, missing tags).
}
\examples{
# tbd
}
