% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReadMe.R
\name{createReadMe}
\alias{createReadMe}
\title{Create a ReadMe file documenting directory contents}
\usage{
createReadMe(
  in_path,
  out_path = NULL,
  lang = c("de", "en"),
  margin = 4,
  col_width = 90,
  indent_per_level = 2,
  max_indent = 10,
  create_table = c("none", "control", "overview", "text"),
  sep = c(";", ","),
  flat_depth = NULL,
  skip_empty_base = FALSE,
  header = NULL,
  content_box = NULL,
  remarks = NULL,
  footer = NULL,
  replace_id = NULL,
  include_rm = NULL
)
}
\arguments{
\item{in_path}{Either a single directory path,
or a \code{character vector} of at least one control table path.}

\item{out_path}{Optional \code{\link{file.path}} (incl. file name) for the ReadMe to be created in.
\link{writeLines} needs to be able to write a plain text ReadMe here.
If \code{NULL} (the defaul), no file is created.}

\item{lang}{Which language(s) should the ReadMe be created in? Per default, German (\code{"de"})
and English (\code{"en"}) are supported, esp. for directory mode (see details). In table mode,
any language acronym that matches a description column in the control tables can be selected.
If more than one language is specified and an \code{out_path} is provided, one ReadMe will be
created for each language, using \code{lang} as the file name suffix.}

\item{margin}{Formatting option: \code{numeric} value of the minimum number of spaces that should
separate file names and file descriptions.
All formatting options only apply if a ReadMe is written to \code{out_path}.}

\item{col_width}{Formatting option: \code{numeric} value of the default width at which the
file descriptions should start. If \code{max(nchar(file names))} > \code{col_width},
width will be set to \code{max(nchar(file names))} and \code{margin} will be applied on that.}

\item{indent_per_level}{Formatting option: \code{numeric} value of the number of spaces used to
indent a line to indicate its contents are one level lower than another line.}

\item{max_indent}{Formatting option: \code{numeric} value of the maximum number
of indentation spaces to use from \code{indent_per_level}.}

\item{create_table}{Optional \code{character} argument to control the return value of this
function (see details).}

\item{sep}{Which column separator is used in the control tables (c.f. \link{read.table})?}

\item{flat_depth}{Directory mode only: \code{numeric} value indicating the depth at which the
function should start pretending that all files and folders are on the same level.
If \code{NULL}, every file's depth will be represented truthfully.
Flattening the depth may prevent file tables from blowing up but you may also consider not
using the function on a very deep directory structure.}

\item{skip_empty_base}{Directory mode only: Should the first level be ignored, if it contains no
direct files? This is useful in creating control tables over many folders at once.}

\item{header, content_box, remarks, footer}{For each of these optional file sections,
either a \code{character vector} of the specific text to be added in the corresponding place
of the ReadMe file, a single valid \code{\link{file.path}} to a control table with the text(s)
in columns named "[irrelevant]_[\code{lang}]" (i.e., "language-coded"), or a list with
"language-coded" names that would result from importing the control table file(s).}

\item{replace_id}{Optional \code{character vector} of length 2 to replace an ID placeholder
(in the FDZ context usually: "_Antrag") with the specific ID (e.g., "2601-01a").
The first element is a \link{regex} expression to be used as \code{pattern}, and replaced by
the simple string in the second element of the vector. Usually only makes sense in table mode.}

\item{include_rm}{Optional \code{logical} argument: Should the ReadMe file be included in itself?
Only applies if a file is created at \code{out_path}. If \code{NULL} (the default), a default
setting depending on the function mode will be applied: \code{FALSE} in directory mode, and
\code{TRUE} in table mode.}
}
\value{
Depending on \code{create_table}, this function may return...\itemize{
 \item \code{"control"} - A \code{list} of \code{data.frame}s, one for each subdirectory.
 These \code{data.frame}s may be saved as individual control table files that could be used
 as input for this function's table mode. Each \code{data.frame} consists of at least three
 columns:
 \code{file_name} (names of individual files and lower subdirectories),
 \code{description} (default descriptions of each file; one column per \code{lang}), and
 \code{flag} (formatting flags to mirror the directory structure if reused in table mode).
 \item \code{"overview"} - A flat \code{data.frame} with at least four columns:
 \code{file_name}, \code{description}, and \code{flag} as described for \code{"control"},
 as well each file's \code{depth} and \code{group} (i.e. first level directory).
 \item \code{"text"} - The \code{character vector} constituting the ReadMe text written to
 \code{out_path} if requested. If more than one \code{lang} is specified, this returns a list
 of each language's ReadMe text.
}
}
\description{
List and describe the contents of a directory including all its subdirectories,
 reflecting the directory structure. Use this function to... \itemize{
  \item create an overview (and, optionally, a control table for future ReadMe construction)
   for an existing directory, or
  \item combine and apply control tables to describe a newly created directory.
 }
}
\details{
This function serves the general purpose of creating a ReadMe that describes the contents of a
 folder/directory. While it can simply be used on an existing directory, its main purpose is in
 creating ReadMes for data packages delivered to the Research Data Centre's data users.
 The function has to modes which are selected based on the content of \code{in_path}:
 \enumerate{
 \item \strong{Directory mode}: If \code{in_path} is a single \code{file.path} that does not
  lead to a (table) file, this function runs in directory mode. It will list all files in the
  supplied directory as well as in all its subdirectories. While this can be used to create a
  simple ReadMe to give an overview of this directory, its main purpose is in creating control
  tables (see the \code{create_table} argument), which may then be enhanced with more specific
  file descriptions and reused by this function in table mode.
 \item \strong{Table mode}: If \code{in_path} is a vector of at least one \code{file.path} that
  leads to a control table file, this function runs in table mode. It will load each control
  table, paste together all tables' contents, and apply formatting to reflect the described
  directory structure. This mode may be used to describe data packages that are themselves
  created from a fixed inventory of data products described in the control tables.
 }
}
\section{Formatting}{
 The look of the ReadMe file can be adjusted by a number of arguments.
 File names and file descriptions are set in a two-column table with a minimum \code{margin}
 between them and with the right column starting no more to the left than at \code{col_width}.
 File names are indented by x = \code{indent_per_level} spaces/blanks under the (sub)headline of
 their respective folder names, which may themselves be indented under their higher-level
 folder's names by the same number of spaces per level, but no more then \code{max_indent}.

 Additional text sections may be added. For table mode, it is recommended to have one control
 table per section, with each table having a column named, e.g., "header_de" for a German header,
 that can be supplied to the function. Currently, the following sections are supported: \itemize{
 \item \code{header} - A centrally aligned header to the whole file, for example stating the name
  of the institution sending out the data package. It is added to the very top of the file.
  Aesthetic lines will be added above and below the header. The line styles are not yet
  customisable, but this functionality may be added on request.
 \item \code{content_box} - A higher-level table of content, further abstracting from the
  detailed file table below. It is added between the header and the main body/file table.
  The contents are indented by one tab and aesthetic lines will be added above and below.
  It may be used to list the "intermediate" data product packages/"studies" of the overall
  data package.
 \item \code{remarks} - A section of additional remarks about using the data. It is added below
  the main body/file table with a simple header above it. The header is not yet customisable,
  but this functionality may be added.
 \item \code{footer} - A footer section added to the very end of the file with decorative lines
  above and below. Use this to present, e.g., contact information.
 }
}

\examples{
# Create dummy directory
somedir <- tempfile("example")
dir.create(somedir)
dir.create(file.path(somedir, "literature"))
dir.create(file.path(somedir, "data"))
file.create(file.path(somedir, "data", c("a_text.txt", "my_data.csv")))
file.create(file.path(somedir, "literature", "Rucker_Weigt_Burblies_Schipolowski_2026.pdf"))

# Direcory mode
createReadMe(somedir, lang = "en", create_table = "overview")

control_table <- createReadMe(somedir, lang = "en", create_table = "control")
control_table
write.table(control_table, file.path(somedir, "ReadMeControl.csv"), sep = ",",
            row.names = FALSE, col.names = c("file_name", "description_en", "flag"))

# Table mode
read.csv(file.path(somedir, "ReadMeControl.csv"))
createReadMe(file.path(somedir, "ReadMeControl.csv"), lang = "en", create_table = "text",
             sep = ",", header = c("This", "is a", "HUGE HEADER"))

# Clean-up
unlink(somedir, recursive = TRUE)


}
