% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fdz.R
\name{download_fdz}
\alias{download_fdz}
\title{Download and import an empty FDZ file.}
\usage{
download_fdz(
  study = c("PISA"),
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_quest")
)
}
\arguments{
\item{study}{Name of the study}

\item{year}{Year of the assessment (only needed for longitudinal assessments).}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty data set from the \href{https://www.iqb.hu-berlin.de/fdz/studies/}{FDZ homepage}.
}
\details{
The function downloads and imports an empty data set (\code{Leerdatensatz}) from the FDZ homepage.
These data sets contain zero rows.
The data is imported via \code{\link[eatGADS]{import_spss}} as a \code{GADSdat} object.
}
\examples{
fdz_pisa <- download_fdz(study = "PISA", year = "2015",
                        data_type = "stud_quest")
}
