% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pisa.R
\name{download_pisa}
\alias{download_pisa}
\title{Download a PISA public use file.}
\usage{
download_pisa(
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_quest")
)
}
\arguments{
\item{year}{Year of the PISA data.}

\item{data_type}{Type of the PISA data.}
}
\description{
Download a PISA public use file from the OECD homepage.
}
\details{
The function downloads a zip file from the OECD homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{haven}.
}
\examples{
# tbd
}
