% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_check_docu.R
\name{reverse_check_docu}
\alias{reverse_check_docu}
\title{Reverse check documentation of data sets.}
\usage{
reverse_check_docu(pdf_path, sav_path, case_sensitive)
}
\arguments{
\item{pdf_path}{Character vector with paths to the \code{.pdf} files.}

\item{sav_path}{Character vector with paths to the \code{.sav} files.}

\item{case_sensitive}{If \code{TRUE}, upper and lower case are differentiated for variable name matching. If \code{FALSE}, case is ignored.}
}
\value{
tbd
}
\description{
short description
}
\details{
more details
}
\examples{
# example r code

}
